priceの型をdecimalへ

+++ database/migrations/2024_06_27_024309_create_menu_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
           $table->unsignedBigInteger('allergies_id');
           $table->unsignedBigInteger('option_id');
           $table->string('menu_image', 255)->nullable();
-          $table->unsignedInteger('price');
+          $table->DECIMAL('price');
           $table->datetime('indate');
           $table->datetime('updated_at');
           $table->foreign('category_id')->references('category_id')->on('categories_table');
